+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/DuratedInputMedia.kt
@@ -1,5 +1,5 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia
 
 interface DuratedInputMedia : InputMedia {
-    val duration: Int?
+    val duration: Long?
 }
